Suport multiple format variables in args and env values
The former restriction on multiple format variables is due
to a bazel limitation on cc_variable substitutions as that's
handled in the java code. Substituting in multiple directories
and files can be handled in pure starlark, so the restriction
on multiple format variables can be limited to just cc_variable
substitutions. This unblocks use cases like formatting the
`INCLUDE` and `LIB` environment variable on windows to specify
the include and library search paths.